OBJ-70 Renaming upload folder
