+++ src/controllers/document_upload/http.request.file.uploader.ts
@@ -10,7 +10,7 @@ import logger from "../../utils/logger";
  * No file data received
  * Upload finished
  */
-export interface UploadFileCallbacks {
+export interface IUploadFileCallbacks {
   /**
    * File size limit exceeded
    * @param {string} filename of the file being uploaded
@@ -37,11 +37,11 @@ export interface UploadFileCallbacks {
  * Looks for the file in a field type "file" in a multipart/form-data form
  * @param {Request} req the http request
  * @param {number} maxFileSizeBytes the maximum allowed size of file in bytes
- * @param {UploadFileCallbacks} callbacks the functions you provide to be executed on file upload events
+ * @param {IUploadFileCallbacks} callbacks the functions you provide to be executed on file upload events
  */
 export const uploadFile = (req: Request,
                            maxFileSizeBytes: number,
-                           callbacks: UploadFileCallbacks) => {
+                           callbacks: IUploadFileCallbacks) => {
 
   const chunkArray: Buffer[] = [];
