+++ src/controllers/document_upload/upload.responder.strategy.factory.ts
