+++ src/controllers/document_upload/upload.responder.strategy.ts
